Calculates comobidites and outcomes from SoS data

Calculates comobidites and outcomes (including time to for the latter) from DIA, OP and ekod variables from National Patient Registry or Cause of Death Registry.

do_sosvar(sosdata, cohortdata, patid = lopnr, indexdate = indexdtm,
  add_unique, sosdate = sosdtm, censdate, type, name,
  starttime = ifelse(type == "out", 1, 0), stoptime, diakod, opkod, ekod,
  diavar = DIA_all, opvar = OP_all, evar = ekod_all,
  meta_reg = "Patientregistret, sluten-, oppenvard- och dagkirurgi",
  meta_pos, warnings = FALSE)

Arguments

sosdata
Data where DIA, OP, Ekod variables are found.

cohortdata
Data containing the cohort with at least columns patid and indexdate.

patid
Patient identifier. Default is lopnr.

indexdate
Index date for the patient (usually date of admission or discharge entered into SwedeHF). Default is indexdtm.

add_unique
If patid and indexdate are not unique in cohortdata, this identifies unique posts.

sosdate
Date of incident (comorbidity or outcome) in sosdata. Default is sosdtm.

censdate
Only for type = "out". Outcomes are up until this time point (usually date of death or censoring).

type
Possible values are "out" and "mh". Is the resulting variable an outcome (and time to is calculated) or medical history (comorbidity)?

name
Name of resulting variable (prefix sos_out_ is added to outcome and sos_mh_ to medical history).

starttime
If type = "out" amount of time (in days) AFTER indexdate to start counted as outcome. If type = "mh" amount of time (in days) PRIOR to indexdate to . Indexdate = 0, all values prior to indexdate are negative. Default is 1 ("out") and 0 ("mh").

stoptime
If type = "out" amount of time (in days) AFTER indexdate to be counted. If type = "mh" amount of time (in days) PRIOR to indexdate to Indexdate = 0, all values prior to indexdate are negative. Default is any time prior to starttime is considered a comorbidity and any time after starttime is considered an outcome.

diakod
String of icd codes as a regular expression defining the outcome/medical history. Should start with blank space " ".

opkod
String of procedure codes as a regular expression defining the outcome/medical history. Should start with blank space " ".

ekod
String of e codes as a regular expression defining the outcome/medical history. Should start with blank space " ".

diavar
Column where diakod is found. All codes should start with blank space " ". Default is DIA_all.

opvar
Column where opkod is found. All codes should start with blank space " ". Default is OP_all.

evar
Column where ekod is found. All codes should start with blank space " ". Default is ekod_all.

meta_reg
Optional argument specifying registries used. Printed in metadatatable. Default is "Patientregistret, sluten-, oppenvard- och dagkirurgi".

meta_pos
Optional argument specifying positions used to search for outcomes/comorbidities. Printed in metadatatable.

warnings
Should warnings be printed. Default is FALSE.

Value

dataset with column containing medical history/outcome. Also dataset metaout that writes directly to global enviroment with information on constructed variables.

See also

Examples

sos_data <- prep_sosdata(sos_data)

rs_data_test <- do_sosvar(
  sosdata = sos_data,
  cohortdata = rs_data,
  patid = id,
  indexdate = indexdtm,
  sosdate = sosdtm,
  type = "mh",
  name = "cv1y",
  diakod = " I",
  stoptime = -365.25
)