Prepares SoS data for input into function do_sosvar

Prepares raw data recieved from Socialstyrelsen (national patient registry or cause of death registry) to be processed by function make_sosvar. Removes post with missing information on relevant dates and joins diagnosis columns together. Reason for preparation is to save space in internal memory (by joining columns together) and removing surplus columns. usethis::use_lifecycle_badge("stable")

prep_sosdata(sosdata, registry = "patreg", utdatum = TRUE,
  remove = TRUE, impute = TRUE)

Arguments

sosdata
Data from Socialstyrelsen that should be prepared. Should contain at least HDIA, OP, ekod, INDATUM

registry
What registry is being prepared? Takes values patreg (National patient registry) or dors (Cause of death registry). Default is National patient registry.

utdatum
Should the sosdtm (usually to be used in comparison with indexdate) be date of discharge for hospitalisation? Default is TRUE. Else date of admission is used.

remove
Should the individual DIA (except HDIA), OP, ekod variables be removed after constructing combined variable (to save space)? Default is TRUE.

impute
To be implemented. kom ihåg att kolla om finns värden i INDATUMA (text) men inte i INDATUM (numerisk). i så fall imputera.

Value

sosdataset with combined DIA, OP, ekod variables and space in front of HDIA and sosdtm.

See also

Examples

# \donttest{
sos_data <- bind_rows(
  sv %>% mutate(source = "SV"),
  ov %>% mutate(source = "OV"),
  dagkir %>% mutate(source = "DK")
)

#> Error in bind_rows(sv %>% mutate(source = "SV"), ov %>% mutate(source = "OV"),     dagkir %>% mutate(source = "DK")): could not find function "bind_rows"

# }
sos_data <- prep_sosdata(sos_data)